v.1.1.2

+++ app/src/main/java/ru/kot1/demo/entity/PostEntity.kt
@@ -14,7 +14,7 @@ data class PostEntity(
     val id: Long,
     val author: String?,
     val authorAvatar: String?,
-    val authorId: Int,
+    val authorId: Long,
     val content: String?,
     @Embedded
     val coords: Coords?,
