+++ app/src/main/java/ru/kot1/demo/entity/forWorker/PostWorkEntity.kt
@@ -14,7 +14,7 @@ data class PostWorkEntity(
     val id: Long,
     val author: String?,
     val authorAvatar: String?,
-    val authorId: Int,
+    val authorId: Long,
     val content: String?,
     @Embedded
     val coords: Coords?,
